portal_streamlit/requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
numpy>=2.0.0
jinja2>=3.1.4
python-dotenv>=1.0.0
